requirements.txt
absl-py
altair==4.2.0
appnope
argon2-cffi
astunparse==1.6.3
async-generator
attrs
backcall
bayesian-optimization==1.2.0
bleach
blinker==1.4
blis==0.7.7
boto3
botocore
Bottleneck
brotlipy==0.7.0
cached-property
cachetools==4.2.4
catalogue==2.0.7
certifi==2021.5.30
cffi
charset-normalizer
click==8.0.3
cloudpickle==2.0.0
cryptography
cycler==0.10.0
cymem==2.0.6
decorator
defusedxml
distlib==0.3.4
docopt==0.6.2
en-core-web-sm
entrypoints==0.3
filelock==3.7.0
Flask
flatbuffers==1.12
funcy==1.17
future==0.18.2
gast
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta
grpcio
gym==0.21.0
h5py 
idna 
imageio
importlib-metadata
inline==0.0.1
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
jedi
Jinja2
jmespath
joblib
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
jupyterlab-pygments
jupyterlab-widgets
kaggle==1.5.12
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing
kiwisolver
langcodes==3.3.0
lxml
Markdown==3.3.4
MarkupSafe
matplotlib
matplotlib-inline
mistune
murmurhash==1.0.7
nb-conda==2.2.1
nb-conda-kernels
nbclient
nbconvert
nbformat
nest-asyncio
nltk==3.7
notebook
numexpr
numpy==1.22.4
oauthlib==3.1.1
olefile
opt-einsum
packaging
pandas==1.3.4
pandas-datareader
pandocfilters
parso
pathy==0.6.1
pexpect
pickleshare
Pillow
pipenv==2022.5.2
pipreqs==0.4.11
platformdirs==2.5.2
praw==7.6.0
prawcore==2.3.0
preshed==3.0.6
prometheus-client
prompt-toolkit
protobuf==3.17.2
psaw==0.1.0
ptyprocess
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
pydantic==1.8.2
pydeck==0.7.1
Pygments 
pyLDAvis==3.3.1
Pympler==1.0.1
pyOpenSSL
pyparsing
PyQt5==5.15.2
PyQt5_sip==4.19.25
PyQtChart==5.15.2
pyrsistent
PySocks
python-dateutil
python-slugify==5.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==20.0.0
qtconsole
QtPy
regex==2022.4.24
requests
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer
scikit-learn
scipy
seaborn==0.11.2
semver==2.13.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
streamlit==1.9.2
termcolor==1.1.0
terminado==0.9.4
text-unidecode==1.3
textblob==0.17.1
thinc==8.0.16
toml==0.10.2
toolz==0.11.2
typer==0.4.1
tzdata==2022.1
tzlocal==4.2
update-checker==0.18.0
urllib3==1.26.7
validators==0.19.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
wasabi==0.9.1
watchdog==2.1.8
webencodings==0.5.1
websocket-client==1.3.2
wordcloud==1.8.1
yarg==0.1.9